+++ packages/core/src/main/Account.ts
@@ -58,6 +58,7 @@ import {
   PayloadBundleType,
 } from './conversation/root';
 import {CryptographyService} from './cryptography/root';
+import {GiphyService} from './giphy/root';
 import {NotificationService} from './notification/root';
 import {SelfService} from './self/root';
 
@@ -65,6 +66,7 @@ import {APIClient} from '@wireapp/api-client';
 import {UserConnectionEvent} from '@wireapp/api-client/dist/commonjs/event';
 import * as EventEmitter from 'events';
 import * as logdown from 'logdown';
+import {UserService} from './user/';
 
 class Account extends EventEmitter {
   private readonly logger = logdown('@wireapp/core/Account', {
@@ -76,11 +78,13 @@ class Account extends EventEmitter {
   public service?: {
     asset: AssetService;
     client: ClientService;
-    conversation: ConversationService;
     connection: ConnectionService;
+    conversation: ConversationService;
     cryptography: CryptographyService;
+    giphy: GiphyService;
     notification: NotificationService;
     self: SelfService;
+    user: UserService;
   };
 
   constructor(apiClient: APIClient = new APIClient()) {
@@ -91,13 +95,16 @@ class Account extends EventEmitter {
   public async init(): Promise<void> {
     this.logger.log('init');
 
+    const assetService = new AssetService(this.apiClient);
     const cryptographyService = new CryptographyService(this.apiClient, this.apiClient.config.store);
+
     const clientService = new ClientService(this.apiClient, this.apiClient.config.store, cryptographyService);
     const connectionService = new ConnectionService(this.apiClient);
-    const assetService = new AssetService(this.apiClient);
+    const giphyService = new GiphyService(this.apiClient);
     const conversationService = new ConversationService(this.apiClient, cryptographyService, assetService);
     const notificationService = new NotificationService(this.apiClient, this.apiClient.config.store);
     const selfService = new SelfService(this.apiClient);
+    const userService = new UserService(this.apiClient);
 
     this.service = {
       asset: assetService,
@@ -105,8 +112,10 @@ class Account extends EventEmitter {
       connection: connectionService,
       conversation: conversationService,
       cryptography: cryptographyService,
+      giphy: giphyService,
       notification: notificationService,
       self: selfService,
+      user: userService,
     };
   }