+++ packages/core/src/main/giphy/GiphyService.ts
@@ -17,10 +17,19 @@
  *
  */
 
-import setGlobalStyles from '../globalStyles';
+import {APIClient} from '@wireapp/api-client';
+import {GiphyOptions, GiphyResult} from '@wireapp/api-client/dist/commonjs/giphy/';
 
-setGlobalStyles();
+class GiphyService {
+  constructor(private readonly apiClient: APIClient) {}
 
-export * from './Logo';
-export * from './colors';
-export * from './Animation';
+  public getRandomGif(tag?: string): Promise<GiphyResult> {
+    return this.apiClient.giphy.api.getGiphyRandom(tag);
+  }
+
+  public searchGif(query: string, options?: GiphyOptions): Promise<GiphyResult> {
+    return this.apiClient.giphy.api.getGiphySearch(query, options);
+  }
+}
+
+export {GiphyService};