+++ packages/core/src/main/user/index.ts
@@ -17,10 +17,4 @@
  *
  */
 
-import setGlobalStyles from '../globalStyles';
-
-setGlobalStyles();
-
-export * from './Logo';
-export * from './colors';
-export * from './Animation';
+export * from './UserService';