+++ packages/react-ui-kit/demo/Demo.jsx
@@ -25,6 +25,7 @@ import {
   AttachmentIcon,
   AudioVideoIcon,
   Bold,
+  BottomUpMovement,
   Box,
   Button,
   ButtonLink,
@@ -60,6 +61,7 @@ import {
   H4,
   HangupIcon,
   HeaderMenu,
+  HeaderSubMenu,
   Heading,
   ICON_NAME,
   ImageIcon,
@@ -72,6 +74,7 @@ import {
   Large,
   Lead,
   LeaveIcon,
+  LeftRightMovement,
   Line,
   Link,
   LinkedInIcon,
@@ -87,6 +90,7 @@ import {
   MoreIcon,
   MuteIcon,
   Muted,
+  Opacity,
   OptionsIcon,
   Overlay,
   PILL_TYPE,
@@ -96,6 +100,7 @@ import {
   PingIcon,
   PlaneIcon,
   ProfileIcon,
+  RightLeftMovement,
   RoundIconButton,
   Select,
   ServicesIcon,
@@ -109,10 +114,13 @@ import {
   TimedIcon,
   Title,
   Tooltip,
+  TopDownMovement,
   TrashIcon,
   TwitterIcon,
   Uppercase,
   WireIcon,
+  XAxisMovement,
+  YAxisMovement,
 } from '@wireapp/react-ui-kit';
 import Color from 'color';
 import Helmet from 'react-helmet';
@@ -164,12 +172,27 @@ ${props.value}${
 
 class Demo extends React.PureComponent {
   state = {
+    animationToggleTimerId: undefined,
     currentPage: 0,
     isFullscreenModalOpen: false,
     isMenuModalOpen: false,
     isModalOpen: false,
     isOverlayOpen: false,
+    showAnimation: false,
   };
+
+  componentDidMount() {
+    const animationInterval = 1000;
+    this.setState({
+      animationToggleTimerId: window.setInterval(() => {
+        this.setState(({showAnimation}) => ({showAnimation: !showAnimation}));
+      }, animationInterval),
+    });
+  }
+
+  componentWillUnmount() {
+    window.clearInterval(this.state.animationToggleTimerId);
+  }
 
   goPage = index => this.setState(state => ({currentPage: index}));
 
@@ -307,8 +330,16 @@ class Demo extends React.PureComponent {
           <MenuLink href="#" color={COLOR.GREEN} button>
             test1
           </MenuLink>
+          <MenuLink href="#">test1</MenuLink>
           <MenuLink href="#">test2</MenuLink>
+          <HeaderSubMenu caption={'Drowdown'}>
+            <MenuLink noWrap>{'Messaging'}</MenuLink>
+            <MenuLink noWrap>{'Voice & video'}</MenuLink>
+            <MenuLink noWrap>{'File sharing & productivity'}</MenuLink>
+          </HeaderSubMenu>
           <MenuLink href="#">test3</MenuLink>
+          <MenuLink href="#">test4</MenuLink>
+          <MenuLink href="#">test5</MenuLink>
         </HeaderMenu>
 
         <Content>
@@ -777,6 +808,56 @@ class Demo extends React.PureComponent {
             <Label>Label</Label>
             <LabelLink block>LabelLink</LabelLink>
             <Line />
+            <H1>Animations</H1>
+            <Columns>
+              <Column>
+                <Opacity in={this.state.showAnimation} startValue={'0'} endValue={'1'}>
+                  <div>Opacity</div>
+                </Opacity>
+              </Column>
+              <Column>
+                <TopDownMovement in={this.state.showAnimation}>
+                  <div>TopDown</div>
+                </TopDownMovement>
+              </Column>
+              <Column>
+                <BottomUpMovement in={this.state.showAnimation}>
+                  <div>BottomUpMovement</div>
+                </BottomUpMovement>
+              </Column>
+              <Column>
+                <YAxisMovement in={this.state.showAnimation} startValue={'50%'} endValue={'-50%'}>
+                  <div>YAxisMovement</div>
+                </YAxisMovement>
+              </Column>
+            </Columns>
+            <br />
+            <br />
+            <Columns>
+              <Column>
+                <LeftRightMovement in={this.state.showAnimation}>
+                  <div>LeftRightMovement</div>
+                </LeftRightMovement>
+              </Column>
+              <Column>
+                <XAxisMovement in={this.state.showAnimation} startValue={'10vh'} endValue={'-10vh'}>
+                  <div>XAxisMovement</div>
+                </XAxisMovement>
+              </Column>
+              <Column>
+                <RightLeftMovement in={this.state.showAnimation}>
+                  <div>RightLeftMovement</div>
+                </RightLeftMovement>
+              </Column>
+            </Columns>
+            <br />
+            <TopDownMovement in={this.state.showAnimation}>
+              <Opacity in={this.state.showAnimation} isInnerAnimation>
+                <XAxisMovement in={this.state.showAnimation} startValue={'40vh'} endValue={'10vh'} isInnerAnimation>
+                  <div>Combined Animation</div>
+                </XAxisMovement>
+              </Opacity>
+            </TopDownMovement>
             <H1>Colors</H1>
             {this.renderColorSection()}
           </Container>