+++ packages/react-ui-kit/CHANGELOG.md
@@ -3,6 +3,51 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="4.2.2"></a>
+## [4.2.2](https://github.com/wireapp/wire-web-packages/tree/master/packages/react-ui-kit/compare/@wireapp/react-ui-kit@4.2.1...@wireapp/react-ui-kit@4.2.2) (2018-10-01)
+
+
+### Bug Fixes
+
+* **react-ui-kit:** Remove no-wrap in the menu link ([#1185](https://github.com/wireapp/wire-web-packages/tree/master/packages/react-ui-kit/issues/1185)) ([d57895b](https://github.com/wireapp/wire-web-packages/tree/master/packages/react-ui-kit/commit/d57895b))
+
+
+
+
+
+<a name="4.2.1"></a>
+## [4.2.1](https://github.com/wireapp/wire-web-packages/tree/master/packages/react-ui-kit/compare/@wireapp/react-ui-kit@4.2.0...@wireapp/react-ui-kit@4.2.1) (2018-10-01)
+
+**Note:** Version bump only for package @wireapp/react-ui-kit
+
+
+
+
+
+<a name="4.2.0"></a>
+# [4.2.0](https://github.com/wireapp/wire-web-packages/tree/master/packages/react-ui-kit/compare/@wireapp/react-ui-kit@4.1.0...@wireapp/react-ui-kit@4.2.0) (2018-10-01)
+
+
+### Features
+
+* **react-ui-kit:** Header Submenu ([#1163](https://github.com/wireapp/wire-web-packages/tree/master/packages/react-ui-kit/issues/1163)) ([d9e7c57](https://github.com/wireapp/wire-web-packages/tree/master/packages/react-ui-kit/commit/d9e7c57))
+
+
+
+
+
+<a name="4.1.0"></a>
+# [4.1.0](https://github.com/wireapp/wire-web-packages/tree/master/packages/react-ui-kit/compare/@wireapp/react-ui-kit@4.0.14...@wireapp/react-ui-kit@4.1.0) (2018-10-01)
+
+
+### Features
+
+* **react-ui-kit:** Animation components ([#1182](https://github.com/wireapp/wire-web-packages/tree/master/packages/react-ui-kit/issues/1182)) ([a221fa1](https://github.com/wireapp/wire-web-packages/tree/master/packages/react-ui-kit/commit/a221fa1))
+
+
+
+
+
 <a name="4.0.14"></a>
 ## [4.0.14](https://github.com/wireapp/wire-web-packages/tree/master/packages/react-ui-kit/compare/@wireapp/react-ui-kit@4.0.13...@wireapp/react-ui-kit@4.0.14) (2018-09-28)